Extend-errors (#15)
This PR extends errors to allow users of the SDK to set the details json
object.

The PR also implements `IntoResponse` for all the error types, so they
can be returned directly from their respective routes.

This means the routes.rs file is removed. Note we add a `fetch_metrics`
function, this contains the logic that used to live in the respective
route (this was the only route with logic).

Before merging, I'll move all the error types into their own file, but
wanted to keep them as-is for now, so it's easier to see what
changed/stayed the same.

You may notice the `IntoResponse` implementation does _not_ use the
`to_string` method on the error types to add the prefix to the message.
This is intentional, that information is conveyed with the status code.