Replace the health check with a readiness check.
The NDC specification states that the health check should respond with
`200 OK` when the connector is ready to start accepting requests (i.e.
"readiness").

Whether the underlying data source is up and running is not important
here.

This adds a default implementation to the method to make it trivial for
connector authors to handle this particular request. The method name has
been changed to clarify matters and also to force connector authors to
understand the change.

We will consider adding further endpoints to check "liveness" and
"connectedness" in the future.