Add a request size limit (#29)
There is a current issue in `sqlx` being investigated:
https://github.com/launchbadge/sqlx/issues/3440

In Engine, we already have request size limits so this isn't an issue,
but no such limits exist in the connectors. This PR adds a 100MB limit
to connector requests, in order to avoid problems if bad actors find
ways to query NDCs directly.