Initialize state on first request. (#31)
We don't need state to successfully handle some requests, so let's not
initialize it until we need it.

With these changes, requests to `/capabilities`, `/schema`, and
`/health` will still succeed even if the state cannot be initialized.

This wraps the state in a `OnceCell` which is initialized on first use.

This changes the `ConnectorSetup` interface, and as such constitutes a
breaking change.