Add a `print-schema-and-capabilities` command (#34)
This adds a `print-schema-and-capabilities` command that will print a
JSON object to stdout that contains the NDC schema and capabilities.
This allows easier access to the connector schema and capabilities
without needing to boot up the webserver and query the `/schema` and
`/capabilities` endpoints.

Example:

```
> cargo run -- print-schema-and-capabilities --configuration .
{"schema":{"scalar_types":{},"object_types":{},"collections":[],"functions":[],"procedures":[]},"capabilities":{"version":"0.1.6","capabilities":{"query":{"nested_fields":{},"exists":{}},"mutation":{}}}}
```

The `print_schema_and_capabilities` is also made public from the crate
so that third parties can use it separately from the default_main
webserver if they want.

---------

Co-authored-by: Benoit Ranque <ranque.benoit@gmail.com>